New AABB class. Implementing AABB collision detection between objects and character. New terrain class.
New AABB collision class with successfully implmenting AABB collision
detection between objects and character. New Terrian class(with
AverageHight improvment).

+++ VS_final_ver/ICT397_Assignment/Camera.cpp
@@ -129,7 +129,7 @@ void Camera::PreventFastDiagMove(){
 	Move(move_vector);
 }
 
-void Camera::isTerrainCollision(int H)
+void Camera::isTerrainCollision(double H)
 {
 	camera_pos.y = H;
 }
